Extract config from fixtures and create locale specific fixtures
