+++ src/Components/HomePage.js
@@ -6,7 +6,7 @@ import theme from '../styles/theme.style';
 import fitness from '../../assets/HomePage/fitness.png';
 import frisbeeGlove from '../../assets/HomePage/frisbeeglove.png';
 import huddle from '../../assets/HomePage/huddle.png';
-import { DrillTypes } from '../Fixtures';
+import { DrillTypes } from '../Fixtures/config';
 import Button from './shared/Button';
 
 export default class HomePage extends React.Component {
