+++ src/Fixtures/config.js
@@ -1,0 +1,57 @@
+export const DrillTypes = {
+  FITNESS: 'fitness',
+  FRISBEE: 'frisbee',
+  GAME: 'game',
+};
+
+export const Intensities = {
+  LOW: 'Low',
+  MODERATE: 'Moderate',
+  HIGH: 'High',
+};
+
+export const FrisbeeGoals = {
+  DEFENSE: 'Defense',
+  MARK: 'Mark',
+  HANDLING: 'Handling',
+  CUTTING: 'Cutting',
+  THROWING: 'Throwing',
+  BREAK: 'Breaking the mark',
+  COMMUNICATION: 'Communication',
+  CATCH: 'Catching',
+  PLAY: 'Playing time',
+};
+
+export const FitnessGoals = {
+  LEGS: 'Legs',
+  UPPER: 'Upper',
+  FULL_BODY: 'Full body',
+  CORE: 'Core',
+  CONDITIONING: 'Conditioning',
+  WARM_UP: 'Warm-up',
+};
+
+export const Levels = {
+  BEGINNER: 'Beginner',
+  INTERMEDIATE: 'Intermediate',
+  ADVANCED: 'Advanced',
+};
+
+export const SeasonTimings = {
+  OFF_SEASON: 'Off-season',
+  PRE_SEASON: 'Pre-season',
+  IN_SEASON: 'In-season',
+  ANYTIME: 'Anytime',
+};
+
+export const EquipmentLabels = {
+  NONE: 'None',
+  BASIC: 'Basic',
+  FULL: 'Full',
+};
+
+export const IllustrationType = {
+  VIMEO: 'vimeo',
+  YOUTUBE: 'youtube',
+  ANIMATION: 'animation',
+};
