+++ src/Fixtures/en.js
@@ -1,0 +1,1872 @@
+import {
+  DrillTypes,
+  Intensities,
+  Levels,
+  FrisbeeGoals,
+  FitnessGoals,
+  SeasonTimings,
+  EquipmentLabels,
+  IllustrationType,
+} from './config';
+
+import animationSquare from './AnimationSquare';
+import animationGoToTen from './AnimationGoToTen';
+
+const stabilityDrill = {
+  id: 1,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Stability Off-season 1',
+  image: 'https://zupimages.net/up/20/19/75s2.jpg',
+  description:
+    "It's the first session of our program, we start with stability drill to build solid base for the rest of the season",
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'Kettlebell - Dumbbell',
+  durationInMinutes: 20,
+  intensity: Intensities.LOW,
+  goals: [FitnessGoals.LEGS, FitnessGoals.UPPER],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Split Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406746924',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 2,
+      title: 'Dumbbell Hip Thrust',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747302',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 3,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 4,
+      title: 'Single Arm Dumbbell Bench Press (Right)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747599',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 5,
+      title: 'Dual Gorilla Kettlebell Rows',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747038',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 6,
+      title: 'Single Leg Vertical Jump (Right)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747769',
+      sounds: false,
+      repetition: '5',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 7,
+      title: 'Split Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406746924',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 8,
+      title: 'Dumbbell Hip Thrust',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747302',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 9,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 10,
+      title: 'Single Arm Dumbbell Bench Press (Left)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747599',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 11,
+      title: 'Dual Gorilla Kettlebell Rows',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747038',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 12,
+      title: 'Single Leg Vertical Jump (Left)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747769',
+      sounds: false,
+      repetition: '5',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 13,
+      title: 'Split Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406746924',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 14,
+      title: 'Dumbbell Hip Thrust',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747302',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 14,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 15,
+      title: 'Single Arm Dumbbell Bench Press (Right)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747599',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 16,
+      title: 'Dual Gorilla Kettlebell Rows',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747038',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 17,
+      title: 'Single Leg Vertical Jump (Right)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747769',
+      sounds: false,
+      repetition: '5',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 18,
+      title: 'Split Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406746924',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 19,
+      title: 'Dumbbell Hip Thrust',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747302',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 20,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 21,
+      title: 'Single Arm Dumbbell Bench Press (Left)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747599',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 22,
+      title: 'Dual Gorilla Kettlebell Rows',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747038',
+      sounds: false,
+      repetition: '12',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 23,
+      title: 'Single Leg Vertical Jump (Left)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747769',
+      sounds: false,
+      repetition: '5',
+      rest: 'As required',
+      instruction: '',
+    },
+  ],
+};
+
+const strengthDrill = {
+  id: 2,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Strenght Off-season 1',
+  image: 'https://zupimages.net/up/20/19/qxbk.jpg',
+  description: 'The first strength practice of the season',
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.FULL,
+  equipment: 'Kettlebell - Dumbbell - Tow bar',
+  durationInMinutes: 15,
+  intensity: Intensities.HIGH,
+  goals: [FitnessGoals.FULL_BODY],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 2,
+      title: 'Single Leg Box Jump (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747656',
+      sounds: false,
+      repetition: '6',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 3,
+      title: 'Single Leg Dumbbell Romanian Deadlift (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747707',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 4,
+      title: 'Pull Ups',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747791',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 5,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 6,
+      title: 'Single Leg Box Jump (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747656',
+      sounds: false,
+      repetition: '6',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 7,
+      title: 'Single Leg Dumbbell Romanian Deadlift (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747707',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 8,
+      title: 'Pull Ups',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747791',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 9,
+      title: 'Goblet Cossack Squats',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747476',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 10,
+      title: 'Single Leg Box Jump (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747656',
+      sounds: false,
+      repetition: '6',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 11,
+      title: 'Single Leg Dumbbell Romanian Deadlift (Alternate)',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747707',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 12,
+      title: 'Pull Ups',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406747791',
+      sounds: false,
+      repetition: '8',
+      rest: '90 s',
+      instruction: '',
+    },
+  ],
+};
+
+const speedDrill = {
+  id: 3,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Speed Off-season 1',
+  image: 'https://zupimages.net/up/20/19/vt6g.jpg',
+  description:
+    "It's the first session of our program, we start with speed drill to build solid base for the rest of the season",
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 30,
+  intensity: Intensities.HIGH,
+  goals: [FitnessGoals.FULL_BODY],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Dynamic Marchs',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406764988',
+      sounds: false,
+      repetition: '2',
+      rest: 'Walk back',
+      instruction: '',
+    },
+    {
+      id: 2,
+      title: 'Wall Fire',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406764961',
+      sounds: false,
+      repetition: '3',
+      rest: 'As required',
+      instruction: '',
+    },
+    {
+      id: 3,
+      title: 'A Skips - 15 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406760709',
+      sounds: false,
+      repetition: '3',
+      rest: 'Walk Back',
+      instruction: '',
+    },
+    {
+      id: 4,
+      title: 'B Skips - 15 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406760726',
+      sounds: false,
+      repetition: '3',
+      rest: 'Walk Back',
+      instruction: '',
+    },
+    {
+      id: 5,
+      title: 'Bounding - 25 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406764918',
+      sounds: false,
+      repetition: '2',
+      rest: 'Walk Back',
+      instruction: '',
+    },
+    {
+      id: 6,
+      title: 'Sprint - 10 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406765010',
+      sounds: false,
+      repetition: '4',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 7,
+      title: 'Sprint - 30 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406765010',
+      sounds: false,
+      repetition: '4',
+      rest: '90 s',
+      instruction: '',
+    },
+    {
+      id: 8,
+      title: 'Sprint - 50 m',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406765010',
+      sounds: false,
+      repetition: '3',
+      rest: '90 s',
+      instruction: '',
+    },
+  ],
+};
+
+const conditioningDrill = {
+  id: 4,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Conditioning Off-season 1',
+  image: 'https://zupimages.net/up/20/19/kj0f.jpg',
+  description: 'Try to follow her for your conditioning',
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 10,
+  intensity: Intensities.HIGH,
+  goals: [FitnessGoals.FULL_BODY],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Follow her',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '406556796',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const warmupGameDrill = {
+  id: 5,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Warm-up Game ready',
+  image: 'https://zupimages.net/up/20/19/f2bd.jpg',
+  description: 'Try to follow him for your warm-up',
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 10,
+  intensity: Intensities.LOW,
+  goals: [FitnessGoals.WARM_UP],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Follow him',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '413628757',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const warmupDrill = {
+  id: 6,
+  type: DrillTypes.FITNESS,
+  author: 'Moby',
+  title: 'Warm-up Ultimate',
+  image: 'https://zupimages.net/up/20/19/va5f.jpg',
+  description: 'Try to follow him for your warm-up',
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FitnessGoals.WARM_UP],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Ultimate warm-up',
+      illustrationType: IllustrationType.VIMEO,
+      sounds: true,
+      illustrationSource: '415430252',
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const conditioningDrillFullBody = {
+  id: 7,
+  type: DrillTypes.FITNESS,
+  author: 'Pamela Reif',
+  title: 'Conditioning Full-body',
+  image: 'https://zupimages.net/up/20/19/kj0f.jpg',
+  description: 'Try to follow her for your conditioning',
+  minimalPlayersNumber: 1,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 20,
+  intensity: Intensities.MODERATE,
+  goals: [FitnessGoals.FULL_BODY],
+  seasonTiming: SeasonTimings.OFF_SEASON,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: 'Follow her',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '416604913',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const basicThrows = {
+  id: 1001,
+  type: DrillTypes.FRISBEE,
+  author: 'Rowan McDonnell',
+  title: 'Basic Throws',
+  image: 'https://zupimages.net/up/20/19/zgxt.jpg',
+  description: 'Basic drills to learn how to throw',
+  minimalPlayersNumber: 2,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc',
+  durationInMinutes: 20,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Backhand Throw',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '415565763',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: 'Follow him',
+    },
+    {
+      id: 2,
+      title: 'Forehand Throw',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '415569048',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: 'Follow him',
+    },
+    {
+      id: 3,
+      title: 'Throwing machine',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'With a partner, the players throw backhands and forehands for 5 minutes. After it, they try to throw in every way possible.',
+    },
+    {
+      id: 4,
+      title: 'The throwing olympics',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'By pair, the players throw as ask by the coach. If the disc falls, they have to go at the right of the 2 lines. The team at the left when the time is over wins!',
+    },
+  ],
+};
+
+const breakDrill = {
+  id: 1023,
+  type: DrillTypes.FRISBEE,
+  author: 'Martin',
+  title: 'Cuts to the break side',
+  image: 'https://zupimages.net/up/20/19/83fx.jpg',
+  description: 'Cut from the open side to the break side. The thrower and the mark are working',
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'Cones if you have some (not mandatory)',
+  durationInMinutes: 20,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.BREAK, FrisbeeGoals.MARK],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '90s',
+      instruction: 'Cut from the open side to the break side. \nChange direction on the fake',
+    },
+    {
+      id: 2,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '90s',
+      instruction: 'The Cuter will do one-step fake and go in the break side',
+    },
+  ],
+};
+
+const threeColumnsLongDrill = {
+  id: 1025,
+  type: DrillTypes.FRISBEE,
+  author: 'Martin',
+  title: '3 columns long shot',
+  image: 'https://zupimages.net/up/20/19/330p.jpg',
+  description: 'Basic long passes drill. You want your players to alternate between throwing, running and passing',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: 'around 1 disc for 3 players, 3 cones',
+  durationInMinutes: 20,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '90s',
+      instruction:
+        "1- The cutter run as fast as possible then he claps the second mate's hand. \n 2- Cut under and long ",
+    },
+  ],
+};
+
+const twoColumnsLongDrill = {
+  id: 1026,
+  type: DrillTypes.FRISBEE,
+  author: 'Martin',
+  title: '2 columns with defense',
+  image: 'https://zupimages.net/up/20/19/fvm4.jpg',
+  description:
+    'Basic long passes drill. You want your players to alternate between throwing, running and passing, with offense and defense',
+  minimalPlayersNumber: 18,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: 'around 1 disc for 3 players, 3 cones',
+  durationInMinutes: 30,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '90s',
+      instruction: '',
+    },
+  ],
+};
+
+const treasureHuntDrill = {
+  id: 1009,
+  type: DrillTypes.FRISBEE,
+  author: 'PUC Ultimate',
+  title: 'Treasure hunt',
+  image: 'https://zupimages.net/up/20/19/z7lq.jpg',
+  description:
+    '2 teams are fighting. The objective is to bring the 5 Frisbees back into a square delimited by the cones.\nPlayers are not allowed to walk with the Frisbee. When a Frisbee falls or is intercepted, it must be brought back to the starting point by the attacking team. The first team that "saves" all its frisbees wins.',
+  minimalPlayersNumber: 10,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '10 discs - 10 cones',
+  durationInMinutes: 20,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.PLAY],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Basic Version',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Goals: Bringing all the disc in the square. \nA player can play in offense on his side or go on the other side to play defense.\nIf there is a turnover, the frisbee goes back to the starting line.\nFirst team that saves all the discs win!',
+    },
+  ],
+};
+
+const kyeDrill = {
+  id: 1010,
+  type: DrillTypes.FRISBEE,
+  author: 'KYE',
+  title: '3 vs 2',
+  image: 'https://zupimages.net/up/20/19/zgxt.jpg',
+  description: 'The team with 2 players stay in defense if they succeed to make the defense',
+  minimalPlayersNumber: 5,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 disc - 2 cones',
+  durationInMinutes: 15,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.DEFENSE, FrisbeeGoals.PLAY, FrisbeeGoals.COMMUNICATION],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Forcing middle\n the offensive cutter can only make two cuts to get the disc.\n1- He try to catch the continuity \n2- He goes under in the middle \nThe goal of this defensive drill is to learn how to quickly set up the mark and how to use your body to control the cut of your opponent. ',
+    },
+  ],
+};
+
+const bigEightDrill = {
+  id: 1011,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'The big 8',
+  image: 'https://zupimages.net/up/20/19/5209.jpg',
+  description: 'Making a pass circuit where each player will try to work on their throws',
+  minimalPlayersNumber: 16,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '8 Discs - 8 Cones',
+  durationInMinutes: 15,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Here is the rotation: \nHandler in the middle \nReset going for the backdoor \nFirst of the stack following in the break \nLast of the stack going long \nSame but in opposite position ',
+    },
+  ],
+};
+
+const trainDrill = {
+  id: 1002,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'The train',
+  image: 'https://zupimages.net/up/20/19/ac8s.jpg',
+  description: 'Drill where you teach the flow with a vertical stack',
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc',
+  durationInMinutes: 10,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.CUTTING, FrisbeeGoals.COMMUNICATION],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: 'Normal train',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Always try to keep the 5 middles and 2 handlers formations. The last of the stack run deep.\nOn the fake, the player come back with an agressive cut in the open side. When he receive the disc all the others players switch their position.\nThe thrower become the reset. The reset become the first of the stack ... and the last of the stack go deep. He has to be running full sprint when the precedent player receive the disc.',
+    },
+    {
+      id: 2,
+      title: 'Racing train',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Keep it simple, and make a race between 2 teams. They have to throw the same amount of passes.',
+    },
+  ],
+};
+
+const goaltimateDrill = {
+  id: 1004,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Goaltimate',
+  image: 'https://zupimages.net/up/20/19/gce7.jpg',
+  description:
+    'The object is to score points by throwing a disc to a teammate in a small scoring area, through a large semicircular hoop called the goal.',
+  minimalPlayersNumber: 8,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 10 Cones or 1 Goal',
+  durationInMinutes: 20,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.DEFENSE, FrisbeeGoals.PLAY, FrisbeeGoals.PLAY],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: 'Goaltimate rules',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        '4 on 4 \nStall count of 5 \nWhenever a turnover happens, the new offensive team must ‘clear’ the disc by taking it into the clear box, and can then attack the goal.\nOne point for a completed pass that travels through the hoop and is caught in the end zone by your teammate',
+    },
+    {
+      id: 2,
+      title: '',
+      illustrationType: IllustrationType.VIMEO,
+      illustrationSource: '416849035',
+      sounds: true,
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const giveAndGoDrill = {
+  id: 1013,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: "Give'n Go",
+  image: 'https://zupimages.net/up/20/19/s9fe.jpg',
+  description: 'Follow your disc as fast as possible',
+  minimalPlayersNumber: 2,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 12 Cones',
+  durationInMinutes: 10,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.HANDLING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Try to do this give and go drill as fast as possible. Be vigilant to the travel.',
+    },
+    {
+      id: 2,
+      title: 'Make it count',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Racing between 2 teams. First to 25 wins!',
+    },
+  ],
+};
+
+const aussieDrill = {
+  id: 1008,
+  type: DrillTypes.FRISBEE,
+  author: 'Matt Hill',
+  title: 'Aussie games',
+  image: 'https://zupimages.net/up/20/19/78g2.png',
+  description: 'You can run with the disc',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '4 Discs - 20 Cones',
+  durationInMinutes: 20,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.COMMUNICATION, FrisbeeGoals.PLAY],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        '2 teams plays against each others.There are 2 territories for each team and they play with 4 discs.\nTo score a point, a player has to bring a disc in one of the endzone of the opponent team. A player can run with a disc. If a player, get touch in the territory of the other team or the disc is intercepted or touch the ground, he looses the possession. In his territory, he is invincible. First team to 5 wins! ',
+    },
+  ],
+};
+
+const bolognaMDrill = {
+  id: 1024,
+  type: DrillTypes.FRISBEE,
+  author: 'Bologna',
+  title: 'The M',
+  image: 'https://zupimages.net/up/20/19/akuy.jpg',
+  description: 'Follow the M',
+  minimalPlayersNumber: 10,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '5 Disc - 6 Cones',
+  durationInMinutes: 20,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.HANDLING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.ADVANCED,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Working on the reset from the first of the stack',
+    },
+  ],
+};
+
+const goToTenDrill = {
+  id: 1014,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Go to 10',
+  image: 'https://zupimages.net/up/20/19/rcyt.jpg',
+  description: 'First team with 10 completes passes win',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 4 Cones',
+  durationInMinutes: 15,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.PLAY, FrisbeeGoals.DEFENSE],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationGoToTen,
+      repetition: '',
+      rest: '',
+      instruction: 'First team with 10 completes passes win. Stall to 6.',
+    },
+  ],
+};
+
+const endzoneDameDrill = {
+  id: 1015,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Endzone Game',
+  image: 'https://zupimages.net/up/20/19/cd36.jpg',
+  description: 'Game to 5. If you score, your team stay in offense',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 4 Cones',
+  durationInMinutes: 15,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.PLAY, FrisbeeGoals.DEFENSE],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: 'Endzone Game',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Game to 5. If you score, your team stay in offense.\nThe player who caught the disc in the endzone has to keep the disc to start the next offense ',
+    },
+    {
+      id: 2,
+      title: 'Endzone finition',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Everybody start behing the disc.\nGame to 5',
+    },
+  ],
+};
+
+const bulldogDrill = {
+  id: 1016,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'British Bulldog',
+  image: 'https://zupimages.net/up/20/19/tf79.jpg',
+  description:
+    'The bulldogs stand in the middle of the play area. All remaining players stand at one end of the area (home). The aim of the game is to run from one end of the field of play to the other, without being caught by the bulldogs. When a player is caught, they become a bulldog themselves.',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '4 Cones',
+  durationInMinutes: 10,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.DEFENSE, FrisbeeGoals.PLAY, FrisbeeGoals.COMMUNICATION],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The bulldogs try to catch the remaining players. The aim of the game is to run from one end of the field of play to the other, without being caught.\nWhen a player is caught, he must hold hands with one another and form a line of 3 bulldogs.\n If there is more than 3 bulldogs, they split in more than one line',
+    },
+  ],
+};
+
+const menageATroisDrill = {
+  id: 1003,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Menage a trois',
+  image: 'https://zupimages.net/up/20/19/ai3z.jpg',
+  description: 'Pass - defense - Receive',
+  minimalPlayersNumber: 3,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc',
+  durationInMinutes: 10,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.DEFENSE, FrisbeeGoals.BREAK],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        "The thrower try to reach the target by using any fake possible (without pivot it's interesting) \nThe defensive player is doing a flat mark",
+    },
+    {
+      id: 2,
+      title: 'Sudden death',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        "The players form 2 columns.\nThe rotation is: Receiving-Throwing-Marking.\nIf player hasn't succeed to complete the throw, he has to make the defense to stay alive. Last 2 players in, win!",
+    },
+  ],
+};
+
+const dwarfDuelDrill = {
+  id: 1012,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Dwarf duel',
+  image: 'https://zupimages.net/up/20/19/3m3k.jpg',
+  description: 'Duel without jumping',
+  minimalPlayersNumber: 3,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 2 Cones',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.CUTTING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        "The coach (or an experience player) is throwing longs.\nWhen the offensive player start to run the defensive player can move. The player can't jump to catch the disc",
+    },
+  ],
+};
+
+const dishieChainDrill = {
+  id: 1017,
+  type: DrillTypes.FRISBEE,
+  author: 'Mooncatcher',
+  title: 'Dishie Chain',
+  image: 'https://zupimages.net/up/20/19/stdq.jpg',
+  description: 'Basic long passes drill. You want your players to alternate between throwing, running and passing',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc - 2 Cones',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.CUTTING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        '2 columns: \n2 players from the same column go in direction of the thrower. When of them receive the disc.\n The thrower is looking ot make a dishie when the other cutter is going long.\n The cutter catching the long throw becomes the new first thrower for the other column',
+    },
+    {
+      id: 2,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Add a defensive player to work on the poach with a side stack',
+    },
+  ],
+};
+
+const breakDanceDrill = {
+  id: 1018,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Break Dance',
+  image: 'https://zupimages.net/up/20/19/agsd.jpg',
+  description: 'Making a christmas tree in the break',
+  minimalPlayersNumber: 12,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '5 Discs - 10 Cones',
+  durationInMinutes: 20,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.CUTTING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.ADVANCED,
+  steps: [
+    {
+      id: 1,
+      title: 'Break continuity',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The cutter try to work on their timing to start their almort horizontal cut. The player goes to the next cones after his throws ',
+    },
+    {
+      id: 2,
+      title: 'The race',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The 2 teams are racing. \nWhen the disc reach the last player, he has to run as fast as possible to start a new serie. \nFirst team that make the complete rotation win!',
+    },
+  ],
+};
+
+const catapultdrill = {
+  id: 1019,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'The Catapult',
+  image: 'https://zupimages.net/up/20/19/kmlc.jpg',
+  description: 'Throwing 10 long shot a row as far and fast as you can',
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '5 Disc - 3 Cones',
+  durationInMinutes: 20,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'One main thrower who is cutting as fast as possible and throwing 10 long throws. \nThe cutter has to find the perfect timing by adapting the start of their cut to the thrower',
+    },
+  ],
+};
+
+const chevronDrill = {
+  id: 1022,
+  type: DrillTypes.FRISBEE,
+  author: 'Author',
+  title: 'Chevron passes',
+  image: 'https://zupimages.net/up/20/19/1y4x.jpg',
+  description: 'Cutting in the right timing, catching and passing straight away',
+  minimalPlayersNumber: 8,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: 'None',
+  durationInMinutes: 10,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.CUTTING, FrisbeeGoals.CATCH, FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const trainDrillHarder = {
+  id: 1020,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'The train 2',
+  image: 'https://zupimages.net/up/20/19/ac8s.jpg',
+  description: 'Drill where you teach the flow with a vertical stack',
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '1 Disc',
+  durationInMinutes: 10,
+  intensity: Intensities.LOW,
+  goals: [FrisbeeGoals.THROWING, FrisbeeGoals.CUTTING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: 'Normal train',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Always try to keep the 5 middles and 2 handlers formations. The last of the stack run deep.\nOn the fake, the player come back with an agressive cut in the open side. When he receive the disc all the others players switch their position.\nThe thrower become the reset. The reset become the first of the stack ... and the last of the stack go deep.\nHe has to be running full sprint when the precedent player receive the disc.',
+    },
+    {
+      id: 2,
+      title: 'With one fake',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Always keep the formation, the thrower can choose to add another fake. If it happens, the cutter clear in the second position of the stack and the reset go up line.',
+    },
+    {
+      id: 3,
+      title: 'With two fakes',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The thrower can choose to make 2 fakes. The reset after going upline clear and the first of the stack go to the reset position',
+    },
+  ],
+};
+
+const resetOpen = {
+  id: 1005,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Reset Open',
+  image: 'https://zupimages.net/up/20/19/tzk3.jpg',
+  description: 'Cutting in the right timing, catching and passing straight away',
+  minimalPlayersNumber: 4,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: '1 Disc',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.CUTTING, FrisbeeGoals.HANDLING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'The backdoor',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The reset handler moves to 45°, in the open side, behind the thrower.\nWhen the thrower makes an eye contact, the reset have to be prepare to receive the backdoor pass.\nThe disc is send back to the thrower, the mark is change and the cut is make again on the other side. ',
+    },
+    {
+      id: 2,
+      title: 'Power reset',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Same position has before. When the eyecontact is made, the disc has to be send in front of the reset handler. Same rotation.',
+    },
+    {
+      id: 3,
+      title: 'Game situation',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Keep the 45°. Now the defense on the reset option is real. The offensive player can choose between the 2 options. No more than one step fake. The team who scores the most wins!',
+    },
+  ],
+};
+
+const resetBreak = {
+  id: 1006,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Reset Break',
+  image: 'https://zupimages.net/up/20/19/w90f.jpg',
+  description: 'Cutting in the right timing, catching and passing straight away',
+  minimalPlayersNumber: 4,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: '1 Disc',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.CUTTING, FrisbeeGoals.HANDLING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'The 45° Up-line',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'The reset handler is on the same line than the thrower.\nThe reset move slowly to a 45° position in the break side. When the thrower makes an eye contact, the reset have to make a incisive up-line cut.\nThe disc is send back to the thrower, the cut is make again on the other side.',
+    },
+    {
+      id: 2,
+      title: 'The easy Reset',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Same start has before. When the eyecontact is made, the disc has to be send close to the starting position of the reset handler. Same rotation.',
+    },
+    {
+      id: 3,
+      title: 'Game situation',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Now the defense on the reset option is real. The offensive player can choose between the 2 options. No more than one step fake. The team who scores the most wins!',
+    },
+  ],
+};
+
+const youShallNotPass = {
+  id: 1007,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'You shall not pass',
+  image: 'https://zupimages.net/up/20/19/90z6.jpg',
+  description: 'Using your body to avoid your opponent to get the disc',
+  minimalPlayersNumber: 4,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: '1 Disc and 4 Cones',
+  durationInMinutes: 12,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.DEFENSE],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Blocking the under',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        "The 2 players will reach the box and then go deep. When he want, the offensive player can come back and try to catch the disc in the box.\nHe can't switch again his direction",
+    },
+    {
+      id: 2,
+      title: 'Blocking everything',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction: 'Same as before but now the offensive player can receive the frisbee deep',
+    },
+    {
+      id: 3,
+      title: 'Make it count',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'One team against another. They score when they catch the disc. Alternate the offense. First team to 10 wins!',
+    },
+  ],
+};
+
+const defensePosition = {
+  id: 1021,
+  type: DrillTypes.FRISBEE,
+  author: 'Moby',
+  title: 'Defense positonning',
+  image: 'https://zupimages.net/up/20/19/zigu.jpg',
+  description: 'Alway keep the good position and adapt to the team defense',
+  minimalPlayersNumber: 4,
+  equipmentLabel: EquipmentLabels.NONE,
+  equipment: '1 Disc and 5 Cones',
+  durationInMinutes: 10,
+  intensity: Intensities.MODERATE,
+  goals: [FrisbeeGoals.DEFENSE],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: 'Blocking the under',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'A thrower with a mark. The offensive cutter can go to any cones.\nThe defensive player have to always keep a distance of 1 meter with the cutter. In this situation, he stays at 45° under.',
+    },
+    {
+      id: 2,
+      title: 'Blocking the deep',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '',
+      rest: '',
+      instruction:
+        'Same as before but the defensive player has to stay at the same level of the cutter. He keeps the 1 meters between him and the cutter',
+    },
+  ],
+};
+
+const normalGame = {
+  id: 2001,
+  type: DrillTypes.GAME,
+  author: 'Martin',
+  title: '7v7 game',
+  image:
+    'https://www.plu.edu/recreations/wp-content/uploads/sites/197/2019/04/ultimate-frisbee-3-2-19-1122-scaled-1536x1163.jpg',
+  description:
+    "It's always interesting to have a time to play in game conditions towards the end of a training. That's the perfect moment to use what players have learnt during the training. Keep in mind that players always FORGET everything when they start playing, so remind them what we're to concentrate on",
+  minimalPlayersNumber: 14,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '8 cones and 1 disc per field',
+  durationInMinutes: 20,
+  intensity: Intensities.LOW,
+  goals: [FitnessGoals.CORE],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const game3v3 = {
+  id: 2002,
+  type: DrillTypes.GAME,
+  author: 'Martin',
+  title: '3v3 games',
+  image:
+    'https://www.plu.edu/recreations/wp-content/uploads/sites/197/2019/04/ultimate-frisbee-3-2-19-1122-scaled-1536x1163.jpg',
+  description:
+    "The perfect setup if you have a LOT of players or want your players to touch a lot of frisbees in a short time. This can be quite exhausting so don't forget to set breaks every 5 o 10 minutes.",
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '8 cones and 1 disc per field',
+  durationInMinutes: 20,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.PLAY],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.BEGINNER,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '',
+      instruction: '',
+    },
+  ],
+};
+
+const longShotGame = {
+  id: 2003,
+  type: DrillTypes.GAME,
+  author: 'Martin',
+  title: 'Long shots oriented game',
+  image:
+    'https://www.plu.edu/recreations/wp-content/uploads/sites/197/2019/04/ultimate-frisbee-3-2-19-1122-scaled-1536x1163.jpg',
+  description:
+    'Just a classical game on the whole field. Only exception: a team scores 2 points if there was a successful long shot on the offense that lead to scoring!',
+  minimalPlayersNumber: 6,
+  equipmentLabel: EquipmentLabels.BASIC,
+  equipment: '8 cones and 1 disc per field',
+  durationInMinutes: 20,
+  intensity: Intensities.HIGH,
+  goals: [FrisbeeGoals.PLAY, FrisbeeGoals.THROWING],
+  seasonTiming: SeasonTimings.ANYTIME,
+  level: Levels.INTERMEDIATE,
+  steps: [
+    {
+      id: 1,
+      title: '',
+      illustrationType: IllustrationType.ANIMATION,
+      illustrationSource: animationSquare,
+      repetition: '1',
+      rest: '',
+      instruction: '2 points if there was a successful long shot on the offense that lead to scoring!',
+    },
+  ],
+};
+
+const dumpSwingTraining = {
+  id: 1,
+  title: 'Dump & Swing',
+  image: 'https://zupimages.net/up/20/19/frj2.jpg',
+  description: 'This training session aims at make your team better at Dump & Swing. blabla bla bla bla',
+  drills: [warmupDrill],
+};
+
+const firstWeekFitnessTraining = {
+  id: 2,
+  title: 'First Week Fitness',
+  image: 'https://zupimages.net/up/20/19/9fcj.jpg',
+  description: 'This is the first week of the Off-season program',
+  drills: [stabilityDrill, strengthDrill, speedDrill, conditioningDrill],
+};
+
+const coed2MarkTraining = {
+  id: 3,
+  title: 'Build your defense',
+  image: 'https://zupimages.net/up/20/19/su0s.jpg',
+  description:
+    "This training session aims at practicing your mark skills. Don't be broken to easily, generate dangerous passes",
+  drills: [warmupGameDrill, dwarfDuelDrill, kyeDrill, breakDrill, normalGame],
+};
+
+const coed2LongTraining = {
+  id: 4,
+  title: 'Long shot is coming',
+  image: 'https://zupimages.net/up/20/19/aapn.jpg',
+  description: 'This training session aims at practicing your long skills.',
+  drills: [warmupGameDrill, threeColumnsLongDrill, longShotGame],
+};
+
+const coed2Long2Training = {
+  id: 5,
+  title: 'Long is your weapon',
+  image: 'https://zupimages.net/up/20/19/v8f4.jpg',
+  description: 'This training session aims at practicing your long skills. ',
+  drills: [warmupDrill, threeColumnsLongDrill, twoColumnsLongDrill, longShotGame],
+};
+
+const firstPracticeChild = {
+  id: 6,
+  title: 'First time',
+  image: 'https://zupimages.net/up/20/19/a7io.png',
+  description: 'Training will allow the players to learn the basics of ultimate while having fun',
+  drills: [warmupDrill, basicThrows, dwarfDuelDrill, game3v3],
+};
+
+const secondPracticeChild = {
+  id: 7,
+  title: 'Defense basics',
+  image: 'https://zupimages.net/up/20/19/dgwo.jpg',
+  description: 'Training will allow the players to learn the basics of ultimate while having fun',
+  drills: [warmupDrill, aussieDrill, menageATroisDrill, game3v3],
+};
+
+const thirdPracticeChild = {
+  id: 8,
+  title: 'Treasure Hunt',
+  image: 'https://zupimages.net/up/20/19/c9i6.jpg',
+  description: 'Training will allow the players to learn the basics of ultimate while having fun',
+  drills: [warmupDrill, bulldogDrill, goToTenDrill, treasureHuntDrill, game3v3],
+};
+
+const secondWeekFitnessTraining = {
+  id: 9,
+  title: 'Second Week Fitness',
+  image: 'https://zupimages.net/up/20/19/9fcj.jpg',
+  description: "This is the second week of the season. It's time to work on your conditioning",
+  drills: [conditioningDrillFullBody, conditioningDrill],
+};
+
+const secondInintiationAdult = {
+  id: 10,
+  title: 'Defense & Longs',
+  image: 'https://zupimages.net/up/20/19/cd36.jpg',
+  description: 'In this training, the players will work on defense and long throws',
+  drills: [warmupDrill, menageATroisDrill, threeColumnsLongDrill, normalGame],
+};
+
+const thirdInintiationAdult = {
+  id: 11,
+  title: 'Adapt to the field',
+  image: 'https://zupimages.net/up/20/19/hrqz.jpg',
+  description: 'In this training, the players will learn to adapt on the size of the field',
+  drills: [warmupDrill, goaltimateDrill, catapultdrill, normalGame],
+};
+
+const initiationVertical = {
+  id: 12,
+  title: 'Vertical Stack',
+  image: 'https://zupimages.net/up/20/19/frj2.jpg',
+  description: 'This training session aims at make your team better at Vertical Stack',
+  drills: [warmupDrill, trainDrill, bigEightDrill, normalGame],
+};
+
+const secondVertical = {
+  id: 13,
+  title: 'The Break power',
+  image: 'https://zupimages.net/up/20/19/frj2.jpg',
+  description: 'This training session aims at make your team better at Vertical Stack',
+  drills: [warmupDrill, resetBreak, trainDrillHarder, game3v3, normalGame],
+};
+
+const defenseBase = {
+  id: 14,
+  title: 'Defense positioning',
+  image: 'https://zupimages.net/up/20/19/skjl.jpg',
+  description: 'This training session aims at practicing your defense.',
+  drills: [warmupGameDrill, defensePosition, youShallNotPass, breakDrill, normalGame],
+};
+
+export default {
+  favoriteDrills: [],
+  completeTrainings: [],
+  drills: [
+    // Fitness
+    stabilityDrill,
+    strengthDrill,
+    speedDrill,
+    conditioningDrill,
+    warmupGameDrill,
+    warmupDrill,
+    conditioningDrillFullBody,
+    // Frisbee
+    basicThrows,
+    trainDrill,
+    menageATroisDrill,
+    goaltimateDrill,
+    resetOpen,
+    resetBreak,
+    youShallNotPass,
+    aussieDrill,
+    treasureHuntDrill,
+    kyeDrill,
+    bigEightDrill,
+    dwarfDuelDrill,
+    giveAndGoDrill,
+    goToTenDrill,
+    endzoneDameDrill,
+    bulldogDrill,
+    dishieChainDrill,
+    breakDanceDrill,
+    catapultdrill,
+    trainDrillHarder,
+    defensePosition,
+    chevronDrill,
+    breakDrill,
+    bolognaMDrill,
+    threeColumnsLongDrill,
+    twoColumnsLongDrill,
+
+    // Games
+    normalGame,
+    game3v3,
+    longShotGame,
+  ],
+  trainings: [
+    dumpSwingTraining,
+    firstWeekFitnessTraining,
+    coed2MarkTraining,
+    coed2LongTraining,
+    coed2Long2Training,
+    firstPracticeChild,
+    secondPracticeChild,
+    thirdPracticeChild,
+    secondWeekFitnessTraining,
+    secondInintiationAdult,
+    thirdInintiationAdult,
+    initiationVertical,
+    secondVertical,
+    defenseBase,
+  ],
+  programs: [
+    {
+      id: 1,
+      title: 'Initiation U13',
+      trainings: [firstPracticeChild, secondPracticeChild, thirdPracticeChild],
+    },
+    {
+      id: 2,
+      title: 'Initiation Adult',
+      trainings: [firstPracticeChild, secondInintiationAdult, thirdInintiationAdult],
+    },
+    {
+      id: 3,
+      title: 'Fitness Bloc',
+      trainings: [firstWeekFitnessTraining, secondWeekFitnessTraining],
+    },
+    {
+      id: 4,
+      title: 'Vertical Stack',
+      trainings: [initiationVertical, secondVertical, dumpSwingTraining],
+    },
+    {
+      id: 5,
+      title: 'Long throws',
+      trainings: [coed2LongTraining, coed2Long2Training],
+    },
+    {
+      id: 6,
+      title: 'Strong defense',
+      trainings: [defenseBase, coed2MarkTraining],
+    },
+  ],
+};
